Add some Markdown cells and clean up code
Markdown gives a better walkthrough of what I'm doing at each step, and some
refactoring makes it easier to follow what is going on (all the plotting stuff
was tossed into a make_plots() subroutine).

Some testing also showed that I needed a contingency plan for when TLAT or
TLONG differed between the two files... I handle that using the baseline values
if the difference is small (and error out if the difference is large).